ai-based-travel-planner-weather-optimized/requirements.txt
streamlit==1.30.0
crewai==0.28.0
crewai-tools==0.3.1
python-dotenv==1.0.0
databricks-sdk